Register htmlbars inline precompile on registry
This refactors the way how the HTMLBarsInlinePrecompilePlugin is
registered on the babel pipeline: previously the ember-cli-babel options
were hijacked and the babel plugin has been added.

Since ember-cli-babel v5.X.X registered addons under the `babel-plugin`
type are added automatically. Since ember-cli-htmlbars v0.7.7 a
`precompile` method is exported on the plugin, so there is no more need
to manually get the HTMLBarsCompiler out of bower_components.