Ensure that nested addons using inline precompile can use Babel 5 locally.
ember-cli-version-checker@1 only looks for top level project dependencies,
this means that if the project is using ember-cli-babel@6 but a nested addon
is still using ember-cli-babel@5 and you add ember-cli-htmlbars-inline-precompile@0.3
to that nested addon it will always throw the error.

This change resolves the local `ember-cli-babel` version, and uses `semver` directly
to check if the version is appropriate. ember-cli-version-checker itself will do the
same check in v2.0.0, but that also updates to ES6 syntax (and therefore would not be
a semver compatible update for this lib).